requirements.txt
Django
python-decouple
djangorestframework
django-rest-framework-mongoengine
markdown
django-filter
selenium
webdriver-manager
pymongo
dnspython
mongoengine
